requirements/base.txt
asgiref==3.8.1
Django==3.2.25
django-filter==2.4.0
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
pytest==8.2.1
pytest-django==4.8.0
